backend/database/pyproject.toml
[project]
name = "alex-database"
version = "0.1.0"
description = "Database package for FinPlex Financial Planner with Aurora Data API"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.8",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
